+++ options.js
@@ -1,15 +1,19 @@
 const defaultOptions = {
   darkMode: false,
-  spawnDelay: 3, // 100*3
+  spawnDelay: 10
 };
 
-let options = { ...defaultOptions, ...JSON.parse(localStorage.getItem("just-some-options") ?? "{}")};
+let options = {
+  ...defaultOptions,
+  ...JSON.parse(localStorage.getItem("just-some-options") ?? "{}")
+};
 
 function updateOption(opt, val) {
   app.options[opt] = val;
   localStorage.setItem("just-some-options", JSON.stringify(options));
   if (opt == "darkMode") {
-    if (val) document.getElementById("theme").innerText = `
+    if (val)
+      document.getElementById("theme").innerText = `
       body, button {
         background-color: #333336 !important;
       }
@@ -25,17 +29,18 @@ function updateOption(opt, val) {
       a:visited {
         color: #ffbbff;
       }
-    `
-    else document.getElementById("theme").innerText = ""
+    `;
+    else document.getElementById("theme").innerText = "";
   }
 }
 
 const app = new Vue({
   el: "#app",
-  data: { options, updateOption },
+  data: { options, updateOption }
 });
 
-if (options.darkMode) document.getElementById("theme").innerText = `
+if (options.darkMode)
+  document.getElementById("theme").innerText = `
   body, button {
     background-color: #333336 !important;
   }
@@ -51,10 +56,10 @@ if (options.darkMode) document.getElementById("theme").innerText = `
   a:visited {
     color: #ffbbff;
   }
-`
+`;
 
 document.getElementById("spawnDelay").value = options.spawnDelay;
 
-document.getElementById("spawnDelay").addEventListener("input", function() {
+document.getElementById("spawnDelay").addEventListener("input", function () {
   updateOption("spawnDelay", this.value);
 });